Refactor break down websocket/mod.rs into sub modules
